combine compiler get_compiler_context and wrapper get_materialization_arguments
